refactor: 패키지 정리 및 사용하지 않는 클래스, import 정리 (#168)
- stock 엔티티 테스트 패키지를 domain -> entity로 변경
- 사용하지 않는 stock facade 삭제
- 기타 사용되지 않는 import 삭제
