--- src/main/java/com/example/i_commerce/domain/product/facade/StockFacade.java
@@ -1,11 +1,0 @@
-package com.example.i_commerce.domain.product.facade;
-
-import com.example.i_commerce.domain.product.application.dto.StockDeductCommand;
-import java.util.List;
-
-public interface StockFacade {
-
-    void deductStock(List<StockDeductCommand> commands);
-    void rollbackStocks(Long orderId);
-
-}
