+++ src/test/java/com/example/i_commerce/domain/product/fixture/ProductItemFixture.java
@@ -3,8 +3,7 @@ package com.example.i_commerce.domain.product.fixture;
 import com.example.i_commerce.domain.product.entity.Product;
 import com.example.i_commerce.domain.product.entity.ProductItem;
 import com.example.i_commerce.domain.product.entity.enums.ProductItemStatus;
-import com.example.i_commerce.domain.product.entity.ProductOptionValue;
-import java.util.List;
+import java.util.ArrayList;
 
 public class ProductItemFixture {
 
@@ -14,7 +13,7 @@ public class ProductItemFixture {
             .sku("SKU-001")
             .price(10000)
             .status(ProductItemStatus.ON_SALE)
-            .attributes(List.of())
+            .attributes(new ArrayList<>())
             .isDefault(true);
 
     }
